Address PR feedback: extract helpers and maintain code organization
- Extract formatRelativeTime to util/dates.ts for reusability
- Extract isNeglected to util/categoryHelpers.ts for category-specific logic
- Keep updatedAt.getTime() call (updatedAt is a Date object from LiveStore)

Addresses feedback from PR #249 review comments.
