+++ packages/web/src/util/dates.ts
@@ -38,3 +38,29 @@ export function formatDateTime(date: Date | string): string {
 export function formatRegistrationDate(date: Date | string): string {
   return formatDateTime(date)
 }
+
+/**
+ * Format a timestamp as relative time
+ * @param timestamp - Timestamp in milliseconds or null
+ * @returns Human-readable relative time (e.g., "Active today", "3 days ago", "2 weeks ago")
+ */
+export function formatRelativeTime(timestamp: number | null): string {
+  if (!timestamp) return 'No activity'
+
+  const now = Date.now()
+  const diffMs = now - timestamp
+  const diffDays = Math.floor(diffMs / (1000 * 60 * 60 * 24))
+  const diffWeeks = Math.floor(diffDays / 7)
+
+  if (diffDays === 0) return 'Active today'
+  if (diffDays === 1) return 'Yesterday'
+  if (diffDays < 7) return `${diffDays} days ago`
+  if (diffWeeks === 1) return '1 week ago'
+  if (diffWeeks < 4) return `${diffWeeks} weeks ago`
+
+  const diffMonths = Math.floor(diffDays / 30)
+  if (diffMonths === 1) return '1 month ago'
+  if (diffMonths < 12) return `${diffMonths} months ago`
+
+  return 'Over a year ago'
+}
